Lecture 56 - Done.
This was a big one as it covered invoke and using states. The scene management may still be a bit basic but going through this stuff was great to do as I barely use invoke, states or enums and I need it drilled into me a hell of a lot more.