Lecture 68 - Done. Lecture 69 - Half Done.
Lecture 68 was just another instructor hangout.

Lecture 69 was to do with loading levels through the build index.
There was a challenge  to do with looping round to the first level after completing the last level.
I couldn't figure it out as it was on something I had never had to do before.
My headache is big. Going to carry on wit this tomorrow.